src/crs/requirements.txt
# --- Web framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# --- HTTP / async ---
httpx>=0.27.0

# --- LLM SDKs (install whichever provider you use) ---
anthropic>=0.39.0
openai>=1.51.0

# --- Retrieval ---
sentence-transformers>=3.1.0
faiss-cpu>=1.9.0
rank-bm25>=0.2.2
numpy>=2.1.0

# --- Data ---
pandas>=2.2.0

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.24.0
